pyproject.toml
[tool.poetry]
name = "discord-reminderbot"
version = "0.1.0"
description = "test"
authors = ["tetsuya-k <64536338+tetsuya-ki@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
async-timeout = "3.0.1"
attrs = "^20.3.0"
chardet = "3.0.4"
"discord.py" = "^1.6.0"
flask = "^1.1.2"
idna = "2.10"
python-dateutil = "^2.8.1"
python-dotenv = "^0.15.0"
six = "^1.15.0"
multidict = "4.7.6"
yarl = "1.5.1"
toml = "^0.10.2"
pycryptodome = "^3.10.1"
discord-py-interactions = "^3.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
